Rename tokentine back to coroutine replacing the implementation
The new implementation seems to be a very good in-place substitute for
the previous version of the coroutine. So just replace it.
