+++ fun/src/coroutine/Coroutine.scala
@@ -6,56 +6,31 @@ import fun.typeclass.Monad
 
 /**
  * General coroutine monad/module. It provides on-demand constant-stack
- * evaluation with client-controlled suspension point.
- *
- * The evaluation model is similar to the `fun.instances.Unnest` monad with
- * the additional ability to "suspend" execution an the given point,
- * and give the client ability to react to such suspension and then resume
- * the rest of the computation.
+ * evaluation with client-controlled "external" evaluation.
  *
  * Instances of this class just capture data and delegate implementation to the
  * static Coroutine (apart from monad instance which explicitly caches result).
  * Although it is possible to just use methods and APIs of the `Coroutine` object,
- * using an instance (capturing Sus type) may be more convenient.
+ * using an instance (capturing Tok type) may be more convenient.
  *
- * @tparam Sus type constructor of the "suspended" execution. This is
- *   client-controlled type that encodes semantics of the suspension (yield value
- *   for generators, "read token" for input/output coroutines, etc...). The type
- *   parameter to Sus encodes return type of the continuation (and the final monad
- *   execution). Coroutines that don't need return value may use trivial type
- *   constructor like `type MySuspension[T] = DiscriminatedUnionType`.
+ * @tparam Tok (super-)type encoding an "external" process.
  */
-final class Coroutine[Sus[_]]:
+final class Coroutine[Tok[_]]:
   /** The type representing routine being executed with coroutine support. */
-  type Routine[T] = Coroutine.Routine[Sus, T]
+  type Routine[T] = Coroutine.Routine[Tok, T]
 
-  /** Result of running coroutine. */
-  type RunResult[T] = Coroutine.RunResult[Sus, T]
-
-  /** Routine suspender. */
-  type Suspender[T] = Coroutine.Suspender[Sus, T]
+  /** Result of running the routine. */
+  type RunResult[T] = Coroutine.RunResult[Tok, T]
 
   /**
-   * Creates a monad instance that suspends execution of the routine and delegates
-   * "coroutine call" to the encoder and then the specific handler. The method
-   * is intended to be called by the specific coroutine implementation and not
-   * the application code.
-   *
-   * This method treats coroutine as a normal value-returning method. Value of
-   * that coroutine call may be used later thus the result of the `suspend` call
-   * is regular value of the monadic type.
-   *
-   * @param encoder encoder for encoding suspension reason and continuation function
-   *   into the format understood by the "coroutine invoker" (code that invoked
-   *   the `run` method).
-   * @return monad that returns the value of the "coroutine" invoked.
+   * Creates a new suspension represented by the given token.
+   * @param token token describing "what" should be done externally to the routine.
    */
-  inline def suspend[T](encoder: Coroutine.Suspender[Sus, T]): Routine[T] =
-    Coroutine.suspend(encoder)
+  inline def suspend[T](token: Tok[T]): Routine[T] =
+    Coroutine.suspend(token)
 
-
-  /** Monad implementation for the coroutine. */
-  given monadInstance: Monad[Routine] = Coroutine.monadInstance[Sus]
+  /** Monad implementation for the routine. */
+  given monadInstance: Monad[Routine] = Coroutine.monadInstance[Tok]
 
 
   /** Runs the monad and returns either the final result or suspension description. */
@@ -64,140 +39,107 @@ final class Coroutine[Sus[_]]:
 end Coroutine
 
 
+/**
+ * A basic coroutine implementation.
+ *
+ * The coroutine use some form of "typed" tokens to encode coroutines and then builds
+ * execution around the concept. The Token[T] encodes some "external process" that will
+ * provide value of the type T.
+ */
 object Coroutine:
   /**
-   * Result of the execution of the suspension.
-   * @tparam Sus type constructor of the suspension information.
-   * @tparam T of the execution.
+   * Result of running some routine with a token-based suspension.
+   * @tparam Tok (super-) type of the token representing an external operation. The
+   *   actual type's argument is the value that is supposed to be returned from the
+   *   corresponding operation.
    */
-  enum RunResult[Sus[_], T]:
+  enum RunResult[Tok[_], T]:
     /**
-     * The execution was suspended.
-     * @param reason reason for which the execution was suspended.
-     */
-    case Suspended(reason: Sus[T])
-    /**
-     * The "routine" completed with the given result.
-     * @param result result of routine evaluation.
+     * Routine evaluation finished with the given outcome.
+     * @param result result of the routine evalutaion.
      */
     case Finished(result: T)
+
+
+    /**
+     * The execution was suspended (with the given token) for an external operation.
+     * @tparam Tok (super-)type of the token representing an external operation.
+     * @tparam Cor specific type that is "returned" by the external operation.
+     * @tparam T type of the whole routine (not the external operation).
+     * @param token token encoding the requested operation.
+     * @param continue function to resume routine evalutaion after value is available
+     *   from some external source.
+     */
+    case Suspended[Tok[_], Cor, T](
+          token: Tok[Cor],
+          continue: Cor => Routine[Tok, T],
+        ) extends RunResult[Tok, T]
   end RunResult
 
 
-  /** Encoding of the computation step(s). */
-  abstract sealed class Routine[Sus[_], +V]
-
-  /** "Pure" value - used as input for next steps. */
-  private case class Pure[Sus[_], +V](value: V) extends Routine[Sus, V]
-
   /**
-   * Suspension - an instruction for the client to perform coroutine-specific step.
-   * The instruction contains suspender (aka factory) because specific continuation
-   * function is not (yet) known. The actual continuation function will be created
-   * during the routine evaluation and will depend on the next operation (if any)
-   * applied to the returned value. I.e. the actual continuation will depend on
-   * `flatMap`s applied to the value returned from the `Continuation.suspend` call.
-   * @tparam Sus suspension encoding type constructor.
-   * @tparam V type of the result returned from the coroutine.
-   * @param suspender encoder used to encode the "suspension reason" to the client.
+   * Routine - something that could be evaluated (with some processes
+   * being executed outside).
+   * @tparam Tok (super-)type used to encode external procesess.
+   * @tparam V type that is produced by the routine.
    */
-  private case class Suspend[Sus[_], V](
-        suspender: Suspender[Sus, V],
-      ) extends Routine[Sus, V]
+  enum Routine[Tok[_], +V]:
+    /** Just a value. */
+    case Pure(value: V)
 
-  /**
-   * Flat map routine.
-   * @param base base computation.
-   * @param fn function to execute on the base value.
-   */
-  private case class FlatMap[Sus[_], V, R](
-        base: Routine[Sus, V],
-        fn: V => Routine[Sus, R],
-      ) extends Routine[Sus, R]
+    /** Standard monadic flatmap operation. */
+    case FlatMap[Tok[_], V, R](
+          base: Routine[Tok, V],
+          fn: V => Routine[Tok, R]
+        ) extends Routine[Tok, R]
+
+    /** Suspension with the given token. */
+    case Suspend[Tok[_], V](token: Tok[V]) extends Routine[Tok, V]
+  end Routine
+
 
 
   /**
-   * Customizable encoding of the suspension point.
-   * The coroutine functionality has to pause the execution, indicate suspension
-   * reason, and provide a way to resume the execution. This trait is responsible
-   * for capturing and encoding buth suspension reason (this is part of suspender
-   * implementation) and the way to resume execution (this is passed to the factory
-   * method).
-   *
-   * @tparam Sus suspension type constructor - the way suspension "reason"
-   *   is encoded for the client.
-   * @tparam T type of the value "returned" from the corresponding "suspendable"
-   *   operation. Value of this type will be available in the "routine" when
-   *   execution is resumed.
+   * Supsends the execution with the given token.
+   * @tparam Tok general (super-)type of all the suspension tokens.
+   * @tparam T type returned by the process encoded by the token.
+   * @param token token describing the "external" process yielding value of type T.
    */
-  trait Suspender[Sus[_], T]:
-    /**
-     * Encodes the suspension reason and "resume" function in the
-     * way the coroutine driver could understand this.
-     * @tparam V type of the "routine" that is being suspended.
-     * @param continue function to resume execution of the "routine".
-     *   The function takes result of the suspension ("everything is an
-     *   expression" so suspension also has some value) and runs until the
-     *   routine is complete or another suspension point occurs.
-     */
-    def encode[V](continue: T => RunResult[Sus, V]): Sus[V]
-  end Suspender
+  inline def suspend[Tok[_], T](token: Tok[T]): Routine[Tok, T] =
+    Routine.Suspend(token)
 
 
-  /**
-   * Creates a monad instance that suspends execution of the routine and delegates
-   * "coroutine call" to the encoder and then the specific handler. The method
-   * is intended to be called by the specific coroutine implementation and not
-   * the application code.
-   *
-   * This method treats coroutine as a normal value-returning method. Value of
-   * that coroutine call may be used later thus the result of the `suspend` call
-   * is regular value of the monadic type.
-   *
-   * @param encoder encoder for encoding suspension reason and continuation function
-   *   into the format understood by the "coroutine invoker" (code that invoked
-   *   the `run` method).
-   * @return monad that returns the value of the "coroutine" invoked.
-   */
-  inline def suspend[Sus[_], T](encoder: Coroutine.Suspender[Sus, T]): Routine[Sus, T] =
-    Suspend(encoder)
+  /** Monad implementation for the routine type. */
+  given monadInstance[Tok[_]]: Monad[({type M[T] = Routine[Tok, T]})#M] with
+    override def pure[T](v: T): Routine[Tok, T] =
+      Routine.Pure(v)
 
-
-  /** Monad instance (factory) for the given Sus type. */
-  given monadInstance[Sus[_]]: Monad[({type M[T] = Routine[Sus, T]})#M] with
-    override def pure[T](v: T): Routine[Sus, T] =
-      Coroutine.Pure(v)
-
-    override def bind[S, R](v: Routine[Sus, S], fn: S => Routine[Sus, R]): Routine[Sus, R] =
-      Coroutine.FlatMap(v, fn)
+    override def bind[S, R](v: Routine[Tok, S], fn: S => Routine[Tok, R]): Routine[Tok, R] =
+      Routine.FlatMap(v, fn)
   end monadInstance
 
 
-
-  /** Runs the monad and returns either the final result or suspension description. */
-  def run[Sus[_], T](routine: Routine[Sus, T]): RunResult[Sus, T] =
+  /** Runs the routine until it completes or until it requests to perform an external operation. */
+  def run[Tok[_], T](routine: Routine[Tok, T]): RunResult[Tok, T] =
     var cur = routine
     while true do
       cur match
-        case Pure(v) => return RunResult.Finished(v)
-        case Suspend(sus) =>
-          /* We have a monad that completely consists of couroutine call. It may be something
-           * stupid (or end result of rewrites) but we have to handle this situation.
-           * Technically, we suspend and when suspension is resolved just return result. I.e.
-           * we use sus.encode(x => run(pure(x))) which after a couple optimization yields
-           * the code below.
+        case Routine.Pure(value) => return RunResult.Finished(value)
+        case Routine.Suspend(token) =>
+          /* The whole routine was just a call to the coroutine.
+           * We still have to encode this into the standard "continuation-based" API.
+           * We do this by re-writing "sus" to "sus.flatMap(pure)" which now could
+           * be expressed via the "RunResult.Suspended"
            */
-          return RunResult.Suspended(sus.encode(RunResult.Finished.apply))
-        case FlatMap(Pure(x), fn) => cur = fn(x)
-        case FlatMap(Suspend(sus), fn) =>
-          /* This one may also be written in a few different ways. For example, we may
-           * encode `x => run(pure(x).flatMap(fn))` but this is just shorter.
-           */
-          return RunResult.Suspended(sus.encode(x => run(fn(x))))
-        case FlatMap(FlatMap(base, fn1), fn) =>
-          cur = FlatMap(base, x => FlatMap(fn1(x), fn))
+          return RunResult.Suspended(token, Routine.Pure.apply)
+        case Routine.FlatMap(Routine.Pure(v), fn) =>  cur = fn(v)
+        case Routine.FlatMap(Routine.Suspend(token), fn) =>
+          return RunResult.Suspended(token, fn)
+        case Routine.FlatMap(Routine.FlatMap(base, fn1), fn) =>
+          cur = Routine.FlatMap(base, x => Routine.FlatMap(fn1(x), fn))
       end match
     end while
     throw new Error("Uncheacheable code reached")
   end run
+
 end Coroutine
